sahayak/pyproject.toml
[project]
name = "sahayak-ai"
version = "2.0.0"
description = "Sahayak AI - Offline Multimodal Medical Assistant for Rural India"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.2",
    "uvicorn[standard]>=0.27.1",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.27",
    "pydantic>=2.6.1",
    "python-dotenv>=1.0.1",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "faiss-cpu>=1.9.0",
    "sentence-transformers>=2.3.1",
    "openai>=1.12.0",
    "groq>=0.9.0",
    "fpdf2>=2.7.8",
    "PyMuPDF>=1.23.0",
    "Pillow>=10.0.0",
    "gTTS>=2.5.1",
    "aiofiles>=23.2.1",
    "numpy>=1.26.4",
    "easyocr>=1.7.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pytesseract>=0.3.10",
]

[project.scripts]
sahayak = "main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["routes", "services", "models", "db", "middleware", "data"]

[dependency-groups]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
